requirements.txt
# Flask and extensions
flask
flask-cors
flask-session
flask-caching

# Core dependencies
python-dotenv
groq
langchain
langchain-groq
langchain-community
sqlalchemy
mysql-connector-python